Continuations are back!
You can now use continuations again, which makes Wee very powerful.
Note that continuations are 100% optional. Simply use Component#callcc
instead of Component#call to call a component.

This commit also refactors the Session so that multiple different
serializers can be used (MutexSerializer, ThreadSerializer). When using
continuations, all requests of a session must be run within the same
thread as continuations cannot be restored from a different thread.

To use continuations simply use the ThreadSerializer:

  Wee.run {
    Wee::Session.new(mycomponent.new, Wee::Session::ThreadSerializer.new)
  }

Note the use of instance variables commented with "CONTINUATIONS!" within
class Session. This is required due to the use of continuations which
replace the stack and as such replace local variables.