Refactor CallbackRegistry to allow to unregister.
Unregistering all callbacks of an object (component) is required when
AJAX is used, where only a part of the page (for example just one
component) rerenders itself and as such registers new callbacks. To
avoid garbage for long running AJAX applications, unregistering the
callbacks before registering new callbacks for the same page is highly
recommended.

The old Array based storage of callbacks was not able to accomplish
that. Using a Hash fixes that.