Refactor action callback invokation
Mehod process_callbacks() now returns the action callback to
invoke. There is only ever one action callback per request.

The advantage is that when continuations are used, the call
stack is relatively flat, as the action callback is invoked
from the session and not from the recursively called
process_callbacks() method. The only drawback is that now
each AnswerDecoration creates an Interceptor object,
which wraps the action callback call to catch "answer"-exceptions.

It is now forbidden to call answer() from an input callback.

While there, optimize the CallbackRegistry a bit.