Refactor root component instanciation
There were issues by unconditionally adding Page and Form decorations
to non-root components which were used as root components (for example
if you write a small component which you don't intend to be used as
a root component of your application but want to quickly test it out),
in that if it adds a PageDecoration itself, it gets wrapped in a
FormDecoration which clearly produces invalid HTML. This can be solved
by changing the add-order of the decorations in Component.instanciate,
but would still produce wrong HTML (duplicate <html> tags etc.).

There is nothing really we can do about it, the only way to do it correctly
is to specify an YourComponentClass.instanciate class method or define
YourComponentClass#root? and return true from it, and add all the decorations
in the correct order in #initialize. For now, only improve the default case,
by checking if one of the decoration has already be added, so that we can
skip them. This can still lead to incorrect HTML, for example if no
FormDecoration is added but a PageDecoration.