Fix problems with too fast AJAX requests
An action request with an action-id without a corresponding
registered action callback is considered an invalid request.

This can happen for AJAX update requests and indicates that
two or more requests are send out too quickly; the first
succeeds, renders and updates a div-tag and registers
a new callback (the old is unregistered). The following
request still uses the old callback id, but now the
callback id has already been unregistered by the previous
request.

In this case we send out a 404 NotFoundResponse.